+++ extensions/application-manager/src/hintTypesRaxInstall/checkHasTypesRax.ts
