+++ extensions/application-manager/src/hintTypesRaxInstall/checkIsRaxTsProject.ts
