+++ extensions/application-manager/src/hintTypesRaxInstall/index.ts
